+++ Dyna/Cards/LockAndLoadCard.cs
@@ -1,9 +1,7 @@
-﻿using daisyowl.text;
+﻿using System.Collections.Generic;
+using System.Reflection;
 using Nanoray.PluginManager;
 using Nickel;
-using Shockah.Kokoro;
-using System.Collections.Generic;
-using System.Reflection;
 
 namespace Shockah.Dyna;
 
@@ -23,86 +21,25 @@ internal sealed class LockAndLoadCard : Card, IRegisterable
 			Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/LockAndLoad.png")).Sprite,
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "LockAndLoad", "name"]).Localize
 		});
-
-		ModEntry.Instance.KokoroApi.CardRendering.RegisterHook(new Hook());
 	}
 
 	public override CardData GetData(State state)
-		=> new()
-		{
-			cost = 2,
-			description = ModEntry.Instance.Localizations.Localize(["card", "LockAndLoad", "description", upgrade.ToString()])
-		};
+		=> new() { cost = 1, description = ModEntry.Instance.Localizations.Localize(["card", "LockAndLoad", "description", upgrade.ToString()]) };
 
 	public override List<CardAction> GetActions(State s, Combat c)
 		=> upgrade switch
 		{
+			Upgrade.A => [
+				new AAddCard { destination = CardDestination.Hand, card = new CustomChargeCard(), amount = 2 },
+				new AStatus { targetPlayer = true, status = Status.evade, statusAmount = 2 },
+			],
 			Upgrade.B => [
-				new AAddCard
-				{
-					destination = CardDestination.Hand,
-					card = new DemoChargeCard
-					{
-						discount = -1,
-						temporaryOverride = true,
-						exhaustOverride = true,
-					},
-					amount = 1
-				},
-				new AAddCard
-				{
-					destination = CardDestination.Hand,
-					card = new FluxChargeCard
-					{
-						discount = -1,
-						temporaryOverride = true,
-						exhaustOverride = true,
-					},
-					amount = 1
-				},
-				new AAddCard
-				{
-					destination = CardDestination.Hand,
-					card = new BurstChargeCard
-					{
-						discount = -1,
-						temporaryOverride = true,
-						exhaustOverride = true,
-					},
-					amount = 1
-				},
-				new AStatus
-				{
-					targetPlayer = true,
-					status = Status.evade,
-					statusAmount = 1
-				}
+				new AAddCard { destination = CardDestination.Hand, card = new CustomChargeCard { upgrade = Upgrade.B }, amount = 1 },
+				new AStatus { targetPlayer = true, status = Status.evade, statusAmount = 1 },
 			],
 			_ => [
-				new AAddCard
-				{
-					destination = CardDestination.Hand,
-					card = new CustomChargeCard(),
-					amount = 2
-				},
-				new AStatus
-				{
-					targetPlayer = true,
-					status = Status.evade,
-					statusAmount = upgrade == Upgrade.A ? 2 : 1
-				}
+				new AAddCard { destination = CardDestination.Hand, card = new CustomChargeCard(), amount = 2 },
+				new AStatus { targetPlayer = true, status = Status.evade, statusAmount = 1 },
 			]
 		};
-
-	private sealed class Hook : IKokoroApi.IV2.ICardRenderingApi.IHook
-	{
-		public Font? ReplaceTextCardFont(IKokoroApi.IV2.ICardRenderingApi.IHook.IReplaceTextCardFontArgs args)
-		{
-			if (args.Card is not LockAndLoadCard)
-				return null;
-			if (args.Card.upgrade != Upgrade.B)
-				return null;
-			return ModEntry.Instance.KokoroApi.Assets.PinchCompactFont;
-		}
-	}
 }